// src/app/api/models/tenant-config.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Address } from '../models/address';

/**
 * Detalles del TenantConfig a crear.
 */
export interface TenantConfig {

  /**
   * Ubicación física o dirección oficial
   */
  address?: Address;

  /**
   * Permite autenticación local
   */
  allowLocal?: boolean;

  /**
   * Capacidad instalada (Camas)
   */
  bedCapacity?: number;

  /**
   * Email de contacto público
   */
  contactEmail?: string;

  /**
   * Teléfono de contacto público
   */
  contactPhone?: string;

  /**
   * Dominio principal del Tenant
   */
  defaultDomain?: string;

  /**
   * Servicio de Emergencia 24/7
   */
  emergency247?: boolean;

  /**
   * Fecha de Expiración del Contrato
   */
  expirationDate?: string;

  /**
   * Configuraciones adicionales (JSON)
   */
  extraJson?: string;

  /**
   * Header personalizado para identificación
   */
  fallbackHeader?: string;

  /**
   * URL del Favicon
   */
  faviconUrl?: string;

  /**
   * Número de Fax
   */
  faxNumber?: string;

  /**
   * Categoría o Nivel
   */
  hospitalCategory?: string;

  /**
   * ID único de configuración
   */
  id?: number;

  /**
   * Razón Social Completa
   */
  legalName?: string;

  /**
   * Configuración regional
   */
  locale?: string;

  /**
   * URL del Logo Principal
   */
  logoUrl?: string;

  /**
   * Límite de usuarios permitidos
   */
  maxUsers?: number;

  /**
   * Nombre del Director Médico
   */
  medicalDirector?: string;

  /**
   * Matrícula del Director
   */
  medicalDirectorLicense?: string;

  /**
   * Licencia Sanitaria General
   */
  medicalLicense?: string;

  /**
   * Fecha de Inicio (Onboarding)
   */
  onboardingDate?: string;

  /**
   * Quirófanos disponibles
   */
  operatingRoomsCount?: number;

  /**
   * Política de contraseñas (JSON)
   */
  passwordPolicyJson?: string;

  /**
   * Habilita verificación de correo
   */
  requireEmailVerification?: boolean;

  /**
   * URL del Sello Institucional
   */
  sealUrl?: string;

  /**
   * Eslogan Institucional
   */
  slogan?: string;

  /**
   * Redes Sociales (JSON)
   */
  socialLinks?: string;

  /**
   * Cuota de almacenamiento en bytes
   */
  storageQuotaBytes?: number;

  /**
   * Número de Identificación Tributaria
   */
  taxId?: string;

  /**
   * Resumen institucional
   */
  tenantDescription?: string;

  /**
   * ID del Tenant asociado
   */
  tenantId: number;

  /**
   * Nombre comercial del Hospital
   */
  tenantName?: string;

  /**
   * Zona horaria operativa
   */
  timezone?: string;

  /**
   * URL del sitio web
   */
  website?: string;

  /**
   * Número oficial de WhatsApp
   */
  whatsappNumber?: string;
}